refactor(sound): extract nested endpoints into separate controller classes
